Add automatic APK open after download option Refactor and clean up code a bit
