+++ library/wversionmanager/src/main/AndroidManifest.xml
@@ -5,6 +5,14 @@
         android:icon="@drawable/ic_launcher"
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
+        <provider
+            android:name="com.winsontan520.wversionmanager.UpdateFileProvider"
+            android:authorities="${applicationId}.UpdateFileProvider"
+            android:exported="false"
+            android:grantUriPermissions="true">
+            <meta-data
+                android:name="android.support.FILE_PROVIDER_PATHS"
+                android:resource="@xml/provider_paths"/>
+        </provider>
     </application>
-
 </manifest>
