+++ library/wversionmanager/src/main/java/com/winsontan520/wversionmanager/UpdateFileProvider.java
@@ -1,0 +1,5 @@
+package com.winsontan520.wversionmanager;
+
+import android.support.v4.content.FileProvider;
+
+public class UpdateFileProvider extends FileProvider { }
